+++ src/components/ui/Footer.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import Link from 'next/link';
 import AnimationContainer from '../utils/AnimationContainer';
 import ExternalLink from './ExternalLink';
@@ -7,8 +9,10 @@ import Credits from '@/src/components/ui/Credits';
 const Footer = () => {
   return (
     <footer className="w-full lg:max-w-screen-md flex flex-col justify-center items-center mx-auto dark:text-white/70 text-black/70">
-      <hr className="w-full border-1 border-black/50 dark:border-white/50 mb-8" />
+      <AnimationContainer customClassName="w-full">
 
+      <hr className="w-full border-1 border-black/50 dark:border-white/50 mb-8" />
+      </AnimationContainer>
       <AnimationContainer customClassName="w-full grid place-items-center grid-cols-3 gap-1 lg:gap-4 pb-8 sm:grid-cols-3 mx-3">
         <div className="flex flex-col space-y-4">
           <Link