+++ src/components/ui/ShowSkills.tsx
@@ -1,10 +1,18 @@
 'use client';
 
 import Image from 'next/image';
+import AnimationContainer from '@/src/components/utils/AnimationContainer';
+import SectionHeader from '@/src/components/ui/SectionHeader';
 
 const MySkills = () => {
 
   return (
+    <AnimationContainer customClassName="w-full mt-16">
+      <div className="flex flex-col gap-5">
+        <SectionHeader
+          title="Skills"
+          content="I’ve been programming for over years, gaining experience with a variety of programming languages, frameworks, and tools. I’ve worked on both Frontend and Backend technologies, allowing me to understand and contribute to the entire development process."
+        />
     <div
       className={` w-full h-full flex justify-center items-center`}
     >
@@ -19,6 +27,8 @@ const MySkills = () => {
 
 
     </div>
+      </div>
+    </AnimationContainer>
   );
 };